requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
aiohttp==3.9.0